kalua: software: opkg_raminstaller() sync with upstream
the 'loader' automatically checks if 'opkg' is installed or not
and set the variable $OPKG according to this. When somebody wants
to install a package and the real 'opkg' is not in place yet it
will be fetched from internet and installed to ramdisk. so a call

$OPKG update
$OPKG install tcpdump-mini

will result in a normal install. when 'opkg' is not built
into the image, this can save ~50 kilobytes. discussed in #197
in another patch i will update the loader to also make this work
when 'opkg' is called directly.